Fix offer expiration timer and display in cart
